requirements.txt
sc-compression
requests
colorama
pylzham
zstandard
pillow
loguru
